refactor: Refactor code in concatenateVideos.js - #1
This commit adds the following changes:

- The 'getFiles' function has been moved to the 'utils' folder to improve code structure.
- Enhanced error handling for a more robust implementation.
- Added cleanup step to delete the 'concat.txt' file containing the list of videos after concatenation.
