--- concat.txt
@@ -1,2 +1,0 @@
-file 'C:/users/aaron/downloads/test/2024-02-10T01-25-03.694Z_segment_1.webm'
-file 'C:/users/aaron/downloads/test/2024-02-10T01-25-03.697Z_segment_2.webm'
